refactor(web): extract Telegram plugin health monitor
Move getClaudePidForSession, hasTelegramPluginAlive, the Marveen 4-stage
recovery escalator (soft -> save -> hard -> gave_up), the per-sub-agent
auto-restart path, hardRestartMarveenChannels, and startTelegramPluginMonitor
into src/web/telegram-monitor.ts. The unused TMUX/CLAUDE locals in web.ts
go away with them.

web.ts: 3616 -> 3311 lines.
